+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed issue where **update-release-notes** added docker image release notes every time the command was called.
 * Fixed an issue where running **update-release-notes** on a pack with newly created integration, had also added a docker image entry in the release notes.
 * Fixed an issue where `XSOAR-linter` did not find *NotImplementedError* in main.
+* Added validation for README files verifying their length (over 30 chars).
 * When using *-g* flag in the **validate** command it will now ignore untracked files by default.
 * Added the *--include-untracked* flag to the **validate** command to include files which are untracked by git in the validation process.